requirements.txt
python-telegram-bot==20.7
fastapi==0.110.2
uvicorn==0.29.0